requirements.txt
numpy>=1.26.0
pandas>=2.1.0
pydicom>=2.4.0
matplotlib>=3.8.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
Pillow>=10.0.0
